+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 * Introduce Sentry to track and report SmileID related errors on the iOS sdk. It is enabled by default using `enableCrashReporting: Bool = true,` on `SmileID.initialize()`
 
+### Changed
+
+* Removed `AntiFraud` response in `JobStatus` calls
+
 ## 11.0.2 - July 10, 2025
 
 ### Changed